+++ Benchmarks/src/Collections/List/ListBenchmarks.cs
@@ -8,8 +8,9 @@ namespace Benchmarks.Collections.List;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class ListBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 	public static readonly List<int> Data = new(1000);
 
@@ -22,7 +23,7 @@ public class ListBenchmarks {
 	[Benchmark("ListGet", "Tests getting values sequentially from a List")]
 	public static int ListGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += Data[CollectionsHelpers.SequentialIndices[j]];
 			}
@@ -34,7 +35,7 @@ public class ListBenchmarks {
 	[Benchmark("ListGet", "Tests getting values randomly from a List")]
 	public static int ListGetRandom() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += Data[CollectionsHelpers.RandomIndices[j]];
 			}
@@ -48,7 +49,7 @@ public class ListBenchmarks {
 		int result = 0;
 		List<int> target = new List<int>();
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int index = 0; index < Data.Count; index++) {
 				target.Add(index);
 			}
@@ -65,7 +66,7 @@ public class ListBenchmarks {
 		int result = 0;
 		List<int> target = new List<int>();
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				target.Add(j);
 			}
@@ -83,7 +84,7 @@ public class ListBenchmarks {
 	[Benchmark("ListCreation", "Tests allocation and initialization of a List")]
 	public static int ListCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			List<int> list = new List<int>();
 			for (int index = 0; index < Data.Count; index++) {
 				list.Add(index * 2);
@@ -98,7 +99,7 @@ public class ListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying a List using a foreach loop")]
 	public static int ListCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			List<int> target = new List<int>();
 			foreach (int element in Data) {
 				target.Add(element);
@@ -113,7 +114,7 @@ public class ListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying a List using a for loop")]
 	public static int ListCopyManualFor() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			List<int> target = new List<int>();
 			for (int index = 0; index < Data.Count; index++) {
 				target.Add(Data[index]);
@@ -129,7 +130,7 @@ public class ListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying a list using a loop where capacity is allocated from the beginning")]
 	public static int ListCopyPreAllocated() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			List<int> target = new List<int>(Data.Count);
 			for (int index = 0; index < Data.Count; index++) {
 				target.Add(index);
@@ -145,7 +146,7 @@ public class ListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying a list using the list constructor")]
 	public static int ListCopyConstructor() {
 		List<int> target = new List<int>();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			target = new List<int>(Data);
 		}
 
@@ -155,7 +156,7 @@ public class ListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying a list using GetRange")]
 	public static int ListCopyGetRange() {
 		List<int> target = new List<int>();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			target = Data.GetRange(0, Data.Count);
 		}
 
@@ -165,7 +166,7 @@ public class ListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying a list using ToList from Linq")]
 	public static int ListCopyManualLinq() {
 		List<int> target = new List<int>();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			target = Data.ToList();
 		}
