+++ Benchmarks/src/HelperObjects/FieldMutabilityHelper.cs
@@ -1,12 +1,12 @@
 ﻿namespace Benchmarks.HelperObjects;
 
 public class FieldMutabilityHelper {
-	public int Field = 10;
-	public const int ConstField = 10;
-	public readonly int ReadonlyField = 10;
+	public ulong Field = 10;
+	public const ulong ConstField = 10;
+	public readonly ulong ReadonlyField = 10;
 
-	public int Property { get; set; } = 10;
-	public int InitProperty { get; init; } = 10;
-	public int GetProperty { get; } = 10;
-	public int ComputedProperty => 10;
+	public ulong Property { get; set; } = 10;
+	public ulong InitProperty { get; init; } = 10;
+	public ulong GetProperty { get; } = 10;
+	public ulong ComputedProperty => 10;
 }
