+++ Benchmarks/src/HelperObjects/Inheritance/InterfaceHelper.cs
@@ -3,10 +3,10 @@ using System.Runtime.CompilerServices;
 namespace Benchmarks.HelperObjects.Inheritance;
 
 public class InterfaceHelper : IDoable {
-	private int _field = 4;
+	private ulong _field = 4;
 
 	[MethodImpl(MethodImplOptions.NoInlining)]
-	public int UpdateAndGetValue() {
+	public ulong UpdateAndGetValue() {
 		_field++;
 		return _field;
 	}
