+++ Benchmarks/src/InheritanceBenchmarks.cs
@@ -7,14 +7,15 @@ namespace Benchmarks;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class InheritanceBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 	[Benchmark("Inheritance", "Tests getting and updating a value with no inheritance")]
-	public static int NoInheritance() {
-		int result = 0;
+	public static ulong NoInheritance() {
+		ulong result = 0;
 		ClassHelper helper = new ClassHelper();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			result += helper.UpdateAndGetValue();
 		}
 
@@ -22,10 +23,10 @@ public class InheritanceBenchmarks {
 	}
 
 	[Benchmark("Inheritance", "Tests getting and updating a value with inheritance")]
-	public static int Inheritance() {
-		int result = 0;
+	public static ulong Inheritance() {
+		ulong result = 0;
 		ClassHelper helper = new InheritanceHelper();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			result += helper.UpdateAndGetValue();
 		}
 
@@ -33,10 +34,10 @@ public class InheritanceBenchmarks {
 	}
 
 	[Benchmark("Inheritance", "Tests getting and updating a value using an polymorphism with an abstract class")]
-	public static int Abstract() {
-		int result = 0;
+	public static ulong Abstract() {
+		ulong result = 0;
 		ADoable helper = new AbstractHelper();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			result += helper.UpdateAndGetValue();
 		}
 
@@ -44,10 +45,10 @@ public class InheritanceBenchmarks {
 	}
 
 	[Benchmark("Inheritance", "Tests getting and updating a value using an polymorphism with a interface")]
-	public static int Interface() {
-		int result = 0;
+	public static ulong Interface() {
+		ulong result = 0;
 		IDoable helper = new InterfaceHelper();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			result += helper.UpdateAndGetValue();
 		}
 
@@ -55,10 +56,10 @@ public class InheritanceBenchmarks {
 	}
 
 	[Benchmark("Inheritance", "Tests getting and updating a value using a virtual method")]
-	public static int InheritanceVirtual() {
-		int result = 0;
+	public static ulong InheritanceVirtual() {
+		ulong result = 0;
 		VirtualHelper helper = new VirtualHelper();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			result += helper.UpdateAndGetValue();
 		}
 
@@ -66,10 +67,10 @@ public class InheritanceBenchmarks {
 	}
 
 	[Benchmark("Inheritance", "Tests getting and updating a value using a overriden virtual method")]
-	public static int InheritanceVirtualOverride() {
-		int result = 0;
+	public static ulong InheritanceVirtualOverride() {
+		ulong result = 0;
 		VirtualHelper helper = new OverrideHelper();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			result += helper.UpdateAndGetValue();
 		}
